Add FreeBSD libc 1.7 abilist files.
Note: I manually removed login_getcapenum from libutil.abilist. This was
necessary because this symbol was added to the library in FreeBSD 14.1.0, which
I based this work on due to 14.0.0 not cleanly building on Linux. collect.zig
normally handles situations like these by checking if the version name is newer
than the one being collected (which is why the abilist files don't contain any
FBSD_1.8 symbols), but libutil is unversioned, so there's no way to detect this
situation.