FreeBSD: Update libc 1.7 abilist files to correct symbol sorting.
Note: I manually removed ftime and login_getcapenum from libutil.abilist. This
was necessary because these symbols were added to the library in FreeBSD 14.2.0
and 14.1.0 respectively. I based this commit on 14.2.0 due to 14.0.0 and 14.1.0
not cleanly building on Linux. collect.zig normally handles situations like
these by checking if the version name is newer than the one being collected
(which is why the abilist files don't contain any FBSD_1.8 symbols), but libutil
is unversioned, so there's no way to detect this situation.